requirements.txt
crewai==0.80.0
langchain-openai==0.2.14
pandas==2.2.3
streamlit==1.41.1
plotly==5.24.1
python-dotenv==1.0.1
rich==13.9.4
tabulate==0.9.0